chore: rename tasks page folder
